refactor project structure
add: List and ListIterator
