+++ src/me/qunwei/iterator/Iterator.java
@@ -1,4 +1,4 @@
-package me.qunwei.define;
+package me.qunwei.iterator;
 
 /**
  * Created by Administrator on 2016/10/20.
